added confirmation modal and fixup member message scroll

+++ Client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -18,7 +19,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading : boolean = false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService : MessageService
+    , private confirmService : ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,9 +37,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m => m.id == id),1);
-    })
+
+    this.confirmService.confirm("Delete message!", "Are you sure to delete this message?").subscribe((result) => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m => m.id == id),1);
+        })
+      }
+    });
+   
   }
 
   pageChanged(event : any){
